added support for time improved AI added scaringEnemy mechanic

+++ Model/StaticHiker.cpp
@@ -3,9 +3,9 @@
 //
 
 #include "StaticHiker.h"
-StaticHiker::StaticHiker(const int &horizontalOffset) {
+StaticHiker::StaticHiker(const float &horizontalOffset) {
     /// todo: 8 divided by lanes doesnt really make sense;
-    Position pos = Position(static_cast<float>(horizontalOffset), 0);
+    Position pos = Position((horizontalOffset), 0);
     Dimentions dim = Dimentions(0.222222f, 0.444444f);
     globalBounds = std::make_shared<GlobalBounds>(pos, dim);
     movementSpeed = 5;
